Change Weight To Double And Change Chartjs Script Tag

+++ Fitness.Models/Goal.cs
@@ -12,7 +12,7 @@ namespace Fitness.Models
         [Key]
         public int Id { get; set; }
         [Required]
-        public int TargetWeight { get; set; }
+        public double TargetWeight { get; set; }
         [AllowedValues(["kgs", "lbs"], ErrorMessage = "Please enter one of 'kgs' or 'lbs'.")]
         [Required]
         public string Unit { get; set; }
